Updates Completed
Folks would like us to promote RR event better on our website. They want to see it on our homepage and at the top of the page in the ribbon if possible.

HomePage - I'm thinking we could add another colour band on the homepage (first ribbon below the YouTube video, above Family Literacy) with a little text summary info and a link to the events page (RR Section):

Text Summary:
Reading Rumble is a book trivia fundraising event for teams to have fun & compete for prizes - all while supporting literacy in Waterloo-Wellington! It is a fun and energizing social event for everyone!  There are many ways to get involved:
Make a team and participate in the trivia tournament!
Sponsor the event or donate prizes!
Please contact info@projectread.ca to get on our mailing list or to discuss other ways that you can support.
(Include button with Link to Reading Rumble section of events page - button could say "Learn More")

Ribbon - Can we also include a new ribbon menu item at the top just called "Reading Rumble" and have that go to the events page (RR Section)

RR Section on Events page - For the Events Page, in light of this focus on the RR event, I would like to re-work the Events page so that we have a separate section now for "Reading Rumble Events", and this could be listed first on the page, above the section for "Community Partnership Events".  This would leave us with just the Portrait Marathon and the Get Set Crop under "Project READ Events" after moving the RR events to the RR section.